refactored linker from service procedure with oo for #45
