+++ dataload/linker/src/main/java/CopyJsonGatheringStringsFromService.java
@@ -1,7 +1,4 @@
 import com.google.gson.JsonElement;
-import com.google.gson.JsonParser;
-import com.google.gson.stream.JsonReader;
-import com.google.gson.stream.JsonToken;
 import com.google.gson.stream.JsonWriter;
 
 import java.io.IOException;
